Adds Python version warning to README
As a brand new Zappa user, I wasted ~2 hours migrating from a local
docker based environment to Virtual Env only to find out Zappa doesn't
support 3.5  :-(.

It looks like others have experienced the same thing too[0], so this PR is
meant to the be a bare minimum notice that Python 3x is not supported.
Down the road, it would be helpful to have a `zappa doctor` type command
to validate/troubleshoot this stuff early on.

Nonetheless, I'm still excited about this project so keep up the good
work!

[0] https://github.com/Miserlou/Zappa/issues/6