New test to remind the user to run doctoc on heading change
This test will fail if a header in README.md changed. It can optionally
write README.test.md, and uses a subset of doctoc's method to generate
a comparable TOC (but does not actually change README.md).

It does not cover all cases, such as `===` style headings, but we use
`###` style in README now anyway.