Fixed typo in issue template
Remove api_gateway policy from tempalte creation

The PR #1903 attempted to fix issue #1747, however, the
travis build failed because at the time `botocore`, required
a version of `docutils` >= 0.10 and <0.15. `botocore` has
since changed the requirements to docutils>=0.10,<0.16. So

This should fix the build issue for the PR to address #1747.

Remove api_gateway policy from tempalte creation

The PR #1903 attempted to fix issue #1747, however, the
travis build failed because at the time `botocore`, required
a version of `docutils` >= 0.10 and <0.15. `botocore` has
since changed the requirements to docutils>=0.10,<0.16. So

This should fix the build issue for the PR to address #1747.

Fix #2003 - Migrate to PEP-518/Poetry

Remove api_gateway policy from tempalte creation

The PR #1903 attempted to fix issue #1747, however, the
travis build failed because at the time `botocore`, required
a version of `docutils` >= 0.10 and <0.15. `botocore` has
since changed the requirements to docutils>=0.10,<0.16. So

This should fix the build issue for the PR to address #1747.