Add basic support for authorization scopes
When using a Cognito User Pool as the authorizer, it is possible to
define which authorization scopes should have access to each API method.

When authorization scopes are defined, the authorizer will use an access
token, whereas the ID token must be used otherwise.

This is only of limited use in the case of Zappa, because it does not
define an API method for every endpoint of the underlying app, but at
least now authorization scopes can be set globally for the entire API.